requirements.txt
streamlit
streamlit-audiorecorder
pydub
audioop-lts
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
